Removing get_for_service_account_* helpers.
After oauth2client 2.0.0 the are just thin wrappers around
factory constructors, so no longer needed.

Note that gcloud.credentials.get_credentials() is also a
very thin wrapper, however, it has useful documentation.
